Move event handling to libeventhandler.
This commit moves all processing of events to a dedicated module.
This includes verify events, pre/post change handlers, constraints/actions, etc.

This extraction is designed to promote modularity, though note
that the mmmk must still be passed to the functions, and many calls
are made on the mmmk directly.
